передал экземпляр authApi, вынес типы, исправил тип 123
