+++ packages/client/src/api/authApiTypes.ts
@@ -1,0 +1,29 @@
+export type SignUpDataType = {
+  first_name: string
+  second_name: string
+  login: string
+  email: string
+  password: string
+  phone: string
+}
+export type SignUpResponse = {
+  id: number
+}
+export type APIError = {
+  reason: string
+}
+
+export type SignInDataType = {
+  login: string
+  password: string
+}
+export type User = {
+  id: string
+  first_name: string
+  second_name: string
+  display_name: string
+  phone: string
+  login: string
+  avatar: string
+  email: string
+}
